pyproject.toml
[project]
name = "libraapi"
version = "0.1.0"
description = ""
authors = [
    {name = "d1vinel3af",email = "d1vinel3af@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "cryptography (>=45.0.3,<46.0.0)",
]

[project.packages]
include = ["libraapi"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
